requirements.txt
pdfplumber>=0.10.0
python-telegram-bot>=20.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
google-genai>=1.0.0
